// projects/ccf-database/src/lib/interfaces.d.ts
export interface SpatialSearch {
    /** X coordinate relative to target in millimeters */
    x: number;
    /** y coordinate relative to target in millimeters */
    y: number;
    /** z coordinate relative to target in millimeters */
    z: number;
    /** Size of the probing sphere in millimeters */
    radius: number;
    /** The target spatial entity IRI */
    target: string;
}
/** Options applied during queries */
export interface Filter {
    /** Sex */
    sex: 'Both' | 'Male' | 'Female';
    /** Age range */
    ageRange: [number, number];
    /** BMI range */
    bmiRange: [number, number];
    /** TMC */
    tmc: string[];
    /** Technologies */
    technologies: string[];
    /** Ontology terms */
    ontologyTerms: string[];
    /** CellType terms */
    cellTypeTerms: string[];
    /** Biomarker terms */
    biomarkerTerms: string[];
    /** spatial search */
    spatialSearches: SpatialSearch[];
    /** Include optional debugging information */
    debug?: boolean;
}